refactor(follow): 패키지 이동 .../common/exception -> .../common/domain/exception
#15
