+++ src/main/java/com/gorogoro/followandlike/follow/domain/exception/FollowException.java
@@ -1,7 +1,6 @@
 package com.gorogoro.followandlike.follow.domain.exception;
 
-import com.gorogoro.followandlike.common.exception.BaseDomainException;
-import lombok.Getter;
+import com.gorogoro.followandlike.common.domain.exception.BaseDomainException;
 
 public class FollowException extends BaseDomainException {
